// src/api/types/document.ts
export type DocumentType = {
    userId: number;
    Index: number;
    TIMESTAMP: string;
    NAME: string;
    CONTACT: string;
    SEX: string;
    'AGE if 12 months 0': string;
    'COUNTRY OF RESIDENCE': string;
    'COUNTY OF RESIDENCE WITHIN KENYA': string;
    'REPORTING COUNTY': string;
    'REPORTING SUB COUNTY': string;
    'REPORTING WARD': string;
    'REPORTING VILLAGE TOWNCAMP': string;
    VACCINATED: string;
    'VACCINE DOSES': string;
    'ATTENDED HEALTH FACILITY YN': string;
    'DATE ATTENDED HEALTH FACILITY ddmmyyyy': string;
    'DATE OF SYMPTOM ONSET ddmmyyyy': string;
    'HEALTH FACILITY NAME lower case only': string;
    'DEHYDRATION AT ARRIVAL': string;
    'RDT RESULT': string;
    'CULTURE RESULT': string;
    'LAB CONFIRMATION DATE ddmmyyyy': string;
    'LAB NAME': string;
    'LAB SITE': string;
    SEROTYPE: string;
    'OUTIN PATIENT STATUS': string;
    OUTCOME: string;
    'DATE OF DISCHARGE ddmmyyyy': string;
    'DATE OF DEATH ddmmyyyy': string;
    'PLACE OF DEATH': string;
    'PREVIOUS HISTORY Epi linked to a confirmed case': string;
    'Comments please provide any other relevant details about the case eg travel history venue symptoms etc': string;
    AgeGroup: string;
    createdAt?: Date;
};

export type DocumentSariType = {
    userId: number;
    PatientID: string;
    Datescreened: string;
    EnrolmentDate: string;
    Eligible: string;
    Enrolled: string;
    F_NAME: string;
    Age: string;
    Sex: string;
    Dateonset: string;
    casedef: string;
    chronic: string;
    flutest: string;
    covidtest: string;
    Dateassesment: string;
    outcome: string;
    colldate: string;
    samplecollected: string;
    sampletested: string;
    datetested: string;
    flupos: string;
    fluapos: string;
    h3n2: string;
    ph1n1: string;
    unsub_non: string;
    notsubtyp: string;
    flubpos: string;
    yamagata: string;
    victoria: string;
    flub_undetermined: string;
    covidpos: string;
    county: string;
    Facility_site: string;
    Longitude: string;
    Latitude: string;
};